app/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.16.0
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==2.1.1
cloudinary==1.41.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1.3
django-allauth==0.61.1
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==22.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.9.2
django-environ==0.9.0
django-filter~=1.1
django-forms-bootstrap==3.1.0
django-rest-swagger==2.2.0
django-storage-url==0.12.1
django-storages==1.14.4
django-webpack-loader==3.1.1
daphne==4.0.0  # Phiên bản phù hợp của Daphne
uwsgi==2.0.21  # Phiên bản phù hợp của uWSGI
django-widget-tweaks==1.4.12
djangocms-admin-style==3.3.1
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.8
ephem==4.1.5
filelock==3.13.1
flake8==7.0.0
frozenlist==1.4.1
furl==2.1.3
geoip2==4.8.0
grpcio==1.60.1
grpcio-status==1.48.2
gspread==6.0.2
gunicorn==23.0.0
httplib2==0.21.0
identify==2.5.35
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
LunarCalendar==0.0.9
MarkupSafe==3.0.2
maxminddb==2.6.2
mccabe==0.7.0
msgpack==1.1.0
multidict==6.0.5
nodeenv==1.8.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==24.0
pandas==2.1.4
Pillow==9.4.0
platformdirs==4.2.0
pre-commit==3.6.2
proto-plus==1.23.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.6.0
PyMySQL==1.1.0
pyparsing==3.0.9
pyshorteners==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0.1
redis==5.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.2.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.4.3
StrEnum==0.4.15
twilio==9.0.5
openpyxl==3.1.5
typing_extensions==4.6.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.25.1
whitenoise==6.7.0
yarl==1.9.4
qrcode==8.0
uvicorn[standard]==0.34.2
google-generativeai==0.8.5